+++ cmd/generate/golang/templates/cobra_program_basic.go.tmpl
@@ -1,0 +1,15 @@
+package golang
+
+import (
+  "fmt"
+
+  "github.com/spf13/cobra"
+)
+
+var {{ .Use }}Cmd = &cobra.Command{
+  Use:   "{{ .Use }}",
+  Short: "{{ .Short }}",
+  Run: func(cmd *cobra.Command, args []string) {
+    fmt.Println("{{ .Use }} called")
+  },
+}
