+++ templates/ansible/group_var.yaml.tmpl
@@ -1,0 +1,4 @@
+---
+{{ range .Values -}}
+{{ .Key }}: "{{ .Value }}"
+{{ end -}}
